Affichage corrigé, dingue ce qui se passe quand on reflechi un peu sur le cours Oo mais ecrase le model quand on rotatione a force

+++ taupe_models/src/Point.java
@@ -3,11 +3,11 @@
  * @author flo
  */
 public class Point {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
     
-    public Point(float x, float y, float z){
+    public Point(double x, double y, double z){
         this.x = x;
         this.y = y;
         this.z = z;
